Move signature schemes to respective files.
