+++ linera-chain/src/certificate/timeout.rs
@@ -2,7 +2,7 @@
 // Copyright (c) Zefchain Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use linera_base::{crypto::Ed25519Signature, data_types::Round, hashed::Hashed};
+use linera_base::{crypto::ed25519::Ed25519Signature, data_types::Round, hashed::Hashed};
 use linera_execution::committee::ValidatorName;
 use serde::{
     ser::{Serialize, SerializeStruct, Serializer},
