+++ linera-core/src/data_types.rs
@@ -6,7 +6,7 @@ use std::{collections::BTreeMap, ops::Not};
 
 use custom_debug_derive::Debug;
 use linera_base::{
-    crypto::{BcsSignable, CryptoError, CryptoHash, Ed25519Signature, KeyPair},
+    crypto::{ed25519::Ed25519Signature, BcsSignable, CryptoError, CryptoHash, KeyPair},
     data_types::{Amount, BlockHeight, Round, Timestamp},
     identifiers::{AccountOwner, ChainDescription, ChainId},
 };
